Reorganize repository into app/infra/ops layout
