+++ apps/common/chat/fool.py
