+++ apps/common/chat/gpt4all.py
