+++ apps/common/stt/whisper.py
