+++ apps/consumer/consumer_requirements.txt
