+++ apps/remetext/__init__.py
@@ -1,0 +1,3 @@
+from apps.remetext.handler import run  # re-export for convenience
+
+__all__ = ["run"]
