+++ apps/webhook/Dockerfile
@@ -2,10 +2,10 @@ FROM python:3.10-slim
 
 WORKDIR /app
 
-COPY requirements.txt .
+COPY apps/webhook/requirements.txt ./requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY server.py .
+COPY apps/webhook/server.py ./server.py
 
 EXPOSE 3000
