+++ infra/ansible/inventory.example.ini
