+++ infra/k8s/remedios.cnf
